Deleted unnecessary classes
