+++ ObjectPrinting/ObjectPrinting.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <LangVersion>8</LangVersion>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
   </PropertyGroup>
