+++ ObjectPrinting/PropertyPrintingConfigExtensions.cs
@@ -1,18 +1,10 @@
-using System;
-
-namespace ObjectPrinting.Solved
+namespace ObjectPrinting
 {
     public static class PropertyPrintingConfigExtensions
-    {
-        public static string PrintToString<T>(this T obj, Func<PrintingConfig<T>, PrintingConfig<T>> config)
-        {
-            return config(ObjectPrinter.For<T>()).PrintToString(obj);
-        }
-
+    { 
         public static PrintingConfig<TOwner> TrimmedToLength<TOwner>(this PropertyPrintingConfig<TOwner, string> propConfig, int maxLen)
         {
             return ((IPropertyPrintingConfig<TOwner, string>)propConfig).ParentConfig;
         }
-
     }
 }
