+++ ObjectPrinting/Tests/ObjectPrinterAcceptanceTests.cs
@@ -1,4 +1,6 @@
-﻿using NUnit.Framework;
+﻿using System;
+using System.Globalization;
+using NUnit.Framework;
 
 namespace ObjectPrinting.Tests
 {
@@ -10,18 +12,21 @@ namespace ObjectPrinting.Tests
         {
             var person = new Person { Name = "Alex", Age = 19 };
 
-            var printer = ObjectPrinter.For<Person>();
-                //1. Исключить из сериализации свойства определенного типа
-                //2. Указать альтернативный способ сериализации для определенного типа
-                //3. Для числовых типов указать культуру
-                //4. Настроить сериализацию конкретного свойства
-                //5. Настроить обрезание строковых свойств (метод должен быть виден только для строковых свойств)
-                //6. Исключить из сериализации конкретного свойства
-            
-            string s1 = printer.PrintToString(person);
+            var printer = ObjectPrinter.For<Person>()
+                .Excluding<Guid>()
+                .Printing<int>().Using(i => i.ToString("X"))
+                .Printing<double>().Using(CultureInfo.InvariantCulture)
+                .Printing(p => p.Name).TrimmedToLength(10)
+                .Excluding(p => p.Age);
 
-            //7. Синтаксический сахар в виде метода расширения, сериализующего по-умолчанию        
-            //8. ...с конфигурированием
+            string s1 = printer.PrintToString(person);
+            
+            string s2 = person.PrintToString();
+            
+            string s3 = person.PrintToString(s => s.Excluding(p => p.Age));
+            Console.WriteLine(s1);
+            Console.WriteLine(s2);
+            Console.WriteLine(s3);
         }
     }
 }
